requirements.txt
aioconsole==0.6.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
mypy==1.1.1
mypy-extensions==1.0.0
openai==0.27.2
pycodestyle==2.10.0
PyQt5-sip==12.11.1
PyQt6==6.4.2
PyQt6-Qt6==6.4.3
PyQt6-sip==13.4.1
pytest==7.3.0
PyYAML==6.0
qasync==0.23.0
requests==2.28.2
Rx==3.2.0
tqdm==4.65.0
types-colorama==0.4.15.8
types-PyYAML==6.0.12.8
types-requests==2.28.11.17
types-urllib3==1.26.25.10
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2